trawler/requirements.txt
websocket-client
python-dotenv
pymongo
requests
pytz
